+++ R/matrix_ops.R
@@ -11,6 +11,10 @@
 #' each column ordered independently from the other columns, of matrix \code{x},
 #' according to the ordering ranks given in matrix \code{mat}. \cr
 #' \cr
+#' In short one can say: \cr
+#'  - \code{x %row~% x} is a much faster version of \code{apply(x, 1, sort)}
+#'  - \code{x %col~% x} is a much faster version of \code{apply(x, 2, sort)} \cr \cr
+#'
 #'
 #' @param x a matrix
 #' @param mat a matrix with the same dimensions as \code{x},
@@ -42,8 +46,9 @@
 #' Re-ordering/sorting every row/column of a matrix with these operators
 #' is generally faster than doing so through loops or apply-like functions. \cr
 #' \cr
-#' Note that these operators strip all attributes except dimensions.
-#'
+#' Note that these operators strip all attributes except dimensions. \cr
+#' \cr
+#' 
 #' @returns
 #' A modified matrix.
 #'