+++ R/tinycodetWrappers.R
@@ -1,0 +1,12 @@
+
+#' @keywords Internal
+#' @noRd
+.C_any_nonpos <- function(x) {
+  .Call("C_any_nonpos", x = x)
+}
+
+#' @keywords Internal
+#' @noRd
+.C_any_neg <- function(x) {
+  .Call("C_any_neg", x = x)
+}