+++ inst/tinytest/regular/test-str_subset_ops.R
@@ -21,7 +21,7 @@ expected.list <- list("abklm", "n", "m", "", x[5], x[6], x[7])
 out.list <- list()
 loops <- loops + 1
 for(i in 1:7) {
-  out.list[[i]] <- x.list[[i]] %sget% as.character(ss[i,])
+  out.list[[i]] <- x.list[[i]] %sget% ss[i,]
   enumerate <- enumerate + 1
 }
 expect_equal(expected.list, out.list)
@@ -31,7 +31,7 @@ expect_equal(x %strim% ss, c("cdefghij", "opqrstuvwxyz", "abcdefghijkl", x[4], r
 expected.list <- list("cdefghij", "opqrstuvwxyz", "abcdefghijkl", x[4], "", "", "")
 loops <- loops + 1
 for(i in 1:7) {
-  out.list[[i]] <- x.list[[i]] %strim% as.character(ss[i,])
+  out.list[[i]] <- x.list[[i]] %strim% ss[i,]
   enumerate <- enumerate + 1
 }
 expect_equal(expected.list, out.list)
@@ -47,6 +47,15 @@ x <- c(
 print(x)
 ss <- rbind(c(2,3), c(1,0), c(0,1), c(0,0))
 expect_error(
+  x %sget% as_chr(ss),
+  pattern = "right hand side must be an integer vector or matrix"
+)
+expect_error(
+  x %strim% as_chr(ss),
+  pattern = "right hand side must be an integer vector or matrix"
+)
+
+expect_error(
   x %sget% -ss,
   pattern = "right hand side cannot contain negative numbers"
 )
@@ -57,12 +66,23 @@ expect_error(
 
 expect_error(
   x %sget% ss[1:2,],
-  pattern = "`nrow(ss)` must be equal to `length(x)`, or must be a vector of length 2",
+  pattern = "right hand side has wrong length or dimensions",
   fixed = TRUE
 )
 expect_error(
   x %strim% ss[1:2,],
-  pattern = "`nrow(ss)` must be equal to `length(x)`, or must be a vector of length 2",
+  pattern = "right hand side has wrong length or dimensions",
+  fixed = TRUE
+)
+
+expect_error(
+  x %sget% cbind(ss, ss),
+  pattern = "right hand side has wrong length or dimensions",
+  fixed = TRUE
+)
+expect_error(
+  x %strim% cbind(ss, ss),
+  pattern = "right hand side has wrong length or dimensions",
   fixed = TRUE
 )