+++ inst/tinytest/regular/test-strfind.R
@@ -194,7 +194,7 @@ rp <- list(
   character(0)
 )
 
-
+loops <- loops + 1
 for(iX in 1:length(x)) {
   for(iP in 1:length(pattern)) {
     for(iRp in 1:length(rp)) {